requirementx.txt
amqp==1.4.8
anyjson==0.3.3
billiard==3.3.0.22
celery==3.1.19
Django==1.8.6
django-celery==3.1.17
djangorestframework==3.3.2
kombu==3.0.30
pytz==2015.7
wheel==0.24.0
